+++ release-notes.md
@@ -1,3 +1,9 @@
+Pending changes in the mainline
+===============================
+
+Fixes:
+  - Prevent using invalid characters when adding labels to a study.
+
 1.4.0 (2024-05-16)
 ==================